RBAC/requirements.txt
django==4.2.11
djangorestframework==3.15.2
django-jinja==2.11.0
psycopg2-binary==2.9.9
